Week 24
Hey @rdzcn! You've been rockin for the last two weeks!

As mentioned during our 1-1, I've tuned up the difficulty of the assignment for this week. It involves a bit of refactoring of the previous work, a bit of polymorphism, separation of concerns, unit testing and setting up a CI system. Please let me know if you hit a blocker.

Keep up the good work! 💪